refactor: split edit.go and client.go along real seams; repair orphaned docs
deslop #103 flagged cmd/wiki/edit.go (781 LOC) and wiki/client.go (530 LOC)
over the 500-LOC threshold, dropping file_health 100% to 90%.

Root cause of edit.go's growth: the -i/--interactive feature (7026dad) added
361 LOC in a single commit, 58% of the final file, appended to the command
file rather than given its own home. The Code-Health-to-10.0 pass (4a9e0ad)
then added a further 156 net LOC of decomposition and doc comments on top.

Seams chosen (not arbitrary 500-line cuts):

* cmd/wiki/edit_interactive.go: the whole -i subsystem. A dedicated test file
  (edit_interactive_test.go) already existed with no production counterpart.
* cmd/wiki/edit_captcha.go: the CAPTCHA retry machinery. Not edit-specific;
  publish.go already reaches into editSession/editRequest.
* readStdin moves to cmd/wiki/client.go. parse.go uses it too, so it was never
  an edit concern; it now sits with the other stdio plumbing.
* wiki/client_request.go: the HTTP request pipeline (rate-limit slot, retry,
  backoff, response classification, decode, redaction), matching the existing
  client_auth.go / client_cache.go / client_helpers.go convention.

Also repairs doc-comment rot introduced by the previous mechanical split
(dd6a3ca, "deslop #59: split 11 files"): 19 doc comments were left behind in
client.go while their functions moved to client_auth.go, client_cache.go and
client_helpers.go. All 16 affected functions, including exported WarmCache,
InvalidateCachePrefix and EnsureLoggedIn, were left undocumented, and
checkExistingSession carried the tail of SetAuditLogger's comment. Comments
restored verbatim from dd6a3ca^.

Pure structural change: zero executable lines added, removed, or altered
(verified line-by-line against origin/main).

go build, go test, go test -race, golangci-lint: all clean.
CodeScene: 10.0 on every file before and after.
