+++ cmd/wiki/client_test.go
@@ -50,3 +50,15 @@ func TestIsQuietMissingFlag(t *testing.T) {
 		t.Error("isQuiet on cmd without quiet flag should be false")
 	}
 }
+
+func TestReadStdinNoPipe(t *testing.T) {
+	// In the test harness stdin is typically a char device (no piped data),
+	// so readStdin returns empty string and no error.
+	got, err := readStdin()
+	if err != nil {
+		t.Fatalf("readStdin: unexpected error: %v", err)
+	}
+	if got != "" {
+		t.Errorf("readStdin with no pipe = %q, want empty", got)
+	}
+}
